+++ src/worker_apis/configuration.rs
@@ -16,7 +16,7 @@ pub struct Configuration {
     pub oauth_access_token: Option<String>,
     pub bearer_access_token: Option<String>,
     pub api_key: Option<ApiKey>,
-    // pub cors: Option<worker::Cors>,
+    pub cors: Option<worker::Cors>,
 }
 
 pub type BasicAuth = (String, Option<String>);
@@ -42,7 +42,7 @@ impl Default for Configuration {
             oauth_access_token: None,
             bearer_access_token: None,
             api_key: None,
-            // cors: None,
+            cors: None,
         }
     }
 }