+++ src/worker_apis/courier_api.rs
@@ -50,7 +50,7 @@ pub async fn get_courier_message(
     // req_builder.set_mode(RequestMode::Cors);
     // req_builder.set_credentials(RequestCredentials::Include);
 
-    let headers = Headers::new();
+    let mut headers = Headers::new();
     // let cors = Cors::new();
 
     if let Some(ref user_agent) = configuration.user_agent {
@@ -67,6 +67,10 @@ pub async fn get_courier_message(
     };
 
     headers.append("Accept", "application/json")?;
+
+    if let Some(ref cors) = configuration.cors {
+        cors.apply_headers(&mut headers)?;
+    }
 
     let mut req_builder = RequestInit::new();
     req_builder.with_method(Method::Get);
@@ -145,7 +149,7 @@ pub async fn list_courier_messages(
         uri_str.add_query(&mut is_first_query, "recipient=", &str);
     }
 
-    let headers = Headers::new();
+    let mut headers = Headers::new();
 
     if let Some(ref user_agent) = configuration.user_agent {
         headers.append("USER_AGENT", user_agent)?;
@@ -161,6 +165,10 @@ pub async fn list_courier_messages(
     };
 
     headers.append("Accept", "application/json")?;
+
+    if let Some(ref cors) = configuration.cors {
+        cors.apply_headers(&mut headers)?;
+    }
 
     let mut req_builder = RequestInit::new();
     req_builder.with_method(Method::Get);