+++ src/worker_apis/metadata_api.rs
@@ -43,13 +43,17 @@ pub async fn get_version(
     configuration: &configuration::Configuration,
 ) -> Result<models::GetVersion200Response, Error<GetVersionError>> {
     let uri_str = format!("{}/version", configuration.base_path);
-    let headers = Headers::new();
+    let mut headers = Headers::new();
 
     if let Some(ref user_agent) = configuration.user_agent {
         headers.append("USER_AGENT", user_agent)?;
     }
 
     headers.append("Accept", "application/json")?;
+
+    if let Some(ref cors) = configuration.cors {
+        cors.apply_headers(&mut headers)?;
+    }
 
     let mut req_builder = RequestInit::new();
     req_builder.with_method(Method::Get);
@@ -100,13 +104,17 @@ pub async fn is_alive(
     configuration: &configuration::Configuration,
 ) -> Result<models::IsAlive200Response, Error<IsAliveError>> {
     let uri_str = format!("{}/health/ready", configuration.base_path);
-    let headers = Headers::new();
+    let mut headers = Headers::new();
 
     if let Some(ref user_agent) = configuration.user_agent {
         headers.append("USER_AGENT", user_agent)?;
     }
 
     headers.append("Accept", "application/json")?;
+
+    if let Some(ref cors) = configuration.cors {
+        cors.apply_headers(&mut headers)?;
+    }
 
     let mut req_builder = RequestInit::new();
     req_builder.with_method(Method::Get);
@@ -157,13 +165,17 @@ pub async fn is_ready(
     configuration: &configuration::Configuration,
 ) -> Result<models::IsAlive200Response, Error<IsReadyError>> {
     let uri_str = format!("{}/health/ready", configuration.base_path);
-    let headers = Headers::new();
+    let mut headers = Headers::new();
 
     if let Some(ref user_agent) = configuration.user_agent {
         headers.append("USER_AGENT", user_agent)?;
     }
 
     headers.append("Accept", "application/json")?;
+
+    if let Some(ref cors) = configuration.cors {
+        cors.apply_headers(&mut headers)?;
+    }
 
     let mut req_builder = RequestInit::new();
     req_builder.with_method(Method::Get);